+++ src/mungers/ca_pri/cdf_elements.txt
@@ -1,7 +1,7 @@
 name	raw_identifier_formula	source
 ReportingUnit	<COUNTY_NAME>	row
 Party	<PARTY_NAME>	row
-CandidateContest	<CONTEST_NAME>	row
+CandidateContest	<CONTEST_NAME> (<PARTY_NAME>)	row
 Candidate	<CANDIDATE_NAME>	row
 BallotMeasureContest	None	row
 BallotMeasureSelection	None	row
