+++ tests/test_dataloader.py
@@ -257,157 +257,3 @@ def test_il_contest_by_vote_type_20(dbname):
 def test_il_totals_match_vote_type_20(dbname):
     assert True == True
 
-
-### California Data Loading Tests ###
-#CA16 test
-
-def test_ca_presidential_16(dbname):
-    assert ( e.contest_total(
-            "2016 General",
-            "California",
-            "US President (CA)",
-            dbname=dbname,
-        )
-            == 14181595
-    )
-
-
-def test_ca_statewide_totals_16(dbname):
-    assert ( e.contest_total(
-            "2016 General",
-            "California",
-            "US Senate CA",
-            dbname=dbname,
-        )
-            == 12244170
-    )
-
-
-def test_ca_senate_totals_16(dbname):
-    assert ( e.contest_total(
-            "2016 General",
-            "California",
-            "CA Senate District 15",
-            dbname=dbname,
-        )
-            == 313531
-    )
-
-
-def test_ca_rep_16(dbname):
-    assert ( e.contest_total(
-            "2016 General",
-            "California",
-            "CA House District 60",
-            dbname=dbname,
-        )
-            == 142114
-    )
-
-
-def test_ca_contest_by_vote_type_16(dbname):
-    assert True == True
-
-
-def test_ca_totals_match_vote_type_16(dbname):
-    assert True == True
-
-
-#CA18 test
-
-def test_ca_presidential_18(dbname):
-    assert True == True
-
-
-def test_ca_statewide_totals_18(dbname):
-    assert ( e.contest_total(
-            "2018 General",
-            "California",
-            "US Senate CA",
-            dbname=dbname,
-        )
-            == 11113364
-    )
-
-
-def test_ca_senate_totals_18(dbname):
-    assert ( e.contest_total(
-            "2018 General",
-            "California",
-            "CA Senate District 12",
-            dbname=dbname,
-        )
-            == 203077
-    )
-
-
-def test_ca_rep_18(dbname):
-    assert ( e.contest_total(
-            "2018 General",
-            "California",
-            "CA House District 60",
-            dbname=dbname,
-        )
-            == 125660
-    )
-
-
-def test_ca_contest_by_vote_type_18(dbname):
-    assert True == True
-
-
-def test_ca_totals_match_vote_type_18(dbname):
-    assert True == True
-
-
-#CA20 test
-
-def test_ca_presidential_20(dbname):
-    assert True == True
-
-
-def test_ca_statewide_totals_20(dbname):
-    assert ( e.contest_total(
-            "2020 Primary",
-            "California",
-            "US Senate CA",
-            dbname=dbname,
-        )
-            == 11113364
-    )
-
-
-def test_ca_senate_totals_20(dbname):
-    assert ( e.contest_total(
-            "2020 Primary",
-            "California",
-            "CA Senate District 12",
-            dbname=dbname,
-        )
-            == 203077
-    )
-
-
-def test_ca_rep_18(dbname):
-    assert ( e.contest_total(
-            "2020 Primary",
-            "California",
-            "CA House District 60",
-            dbname=dbname,
-        )
-            == 125660
-    )
-
-
-
-
-def test_ca_presidential_20ppp(dbname):
-    assert ( e.contest_total(
-            "2020 President Preference Primary",
-            "California",
-            "US President (CA)",
-            dbname=dbname,
-        )
-            == 2780247
-    )
-
